fix type

+++ server/internal/domain/service/getIdFromJWT.go
@@ -5,11 +5,11 @@ import (
 	"github.com/golang-jwt/jwt/v4"
 )
 
-func GetUserIDFromJWT(c *fiber.Ctx) (int, error) {
+func GetUserIDFromJWT(c *fiber.Ctx) (float64, error) {
 	user := c.Locals("user")
 	token := user.(*jwt.Token)
 	claims := token.Claims.(jwt.MapClaims)
-	userID := claims["user_id"].(int)
+	userID := claims["user_id"].(float64)
 
 	return userID, nil
 }
